Attempt to fix an issue with unlinking files that do not exist.

+++ app/Http/Controllers/RecipeController.php
@@ -88,7 +88,7 @@ class RecipeController extends Controller
             $request->request->add(['image' => $imageFile]);
             $updates[] = 'image';
 
-            if (!is_null($recipe->image)) {
+            if (!is_null($recipe->image) && file_exists('app/public/recipe_images/' . end($oldImageParts))) {
                 $oldImageParts = explode('/', $recipe->image);
                 $fileToDelete = storage_path('app/public/recipe_images/' . end($oldImageParts));
                 unlink($fileToDelete);
